Initial eval support
The eval result is converted to JSON and then the string will be passed
to the caller in the result strucuture. This was the easiest way to get
this going without having to resort to translating all the different
types and C++-ishms to C and further down the road to Rust.